enable ct_cleanup at the start of the run script
Reason:
The ct_enable_cleanup only set trap builtin to call ct_cleanup on EXIT
or on SIGINT.
From manpages:
If a sigspec is EXIT (0) the command arg is executed on exit from  the
shell.
That would mean, that even though the ct_enable_cleanup was called in a
specific function before, its effect would only appear when exiting from
shell (when the tests are terminated) (or on SIGINT). Thus the
ct_enable_cleanup has never been used for that function's scope.